Separated the change password and change profile forms into their own files
